requirements.txt
asttokens==3.0.0
certifi==2025.8.3
charset-normalizer==3.4.2
comm==0.2.3
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.15
decorator==5.2.1
exceptiongroup==1.3.0
executing==2.2.0
fonttools==4.59.0
idna==3.10
ipykernel==6.30.1
ipython==8.37.0
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.8
matplotlib==3.10.5
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==1.24.4
packaging==25.0
pandas==2.0.3
parso==0.8.4
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pyzmq==27.0.1
requests==2.28.2
six==1.17.0
stack-data==0.6.3
tornado==6.5.1
traitlets==5.14.3
typing_extensions==4.14.1
tzdata==2025.2
urllib3==1.26.20
wcwidth==0.2.13
